MGDCTRS-568: Move connector actions menu to a component
